@@ -3,60 +3,65 @@
   "version": "1.0.0-alpha.2",
   "private": true,
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start",
-    "test": "jest"
+    "android": "react-native run-android",
+    "ios": "react-native run-ios",
+    "start": "react-native start",
+    "test": "jest",
+    "lint": "eslint .",
+    "build:android": "cd android && ./gradlew assembleRelease"
   },
   "dependencies": {
-    "axios": "^0.18.1",
-    "formik": "^1.5.4",
+    "axios": "^0.19.0",
+    "formik": "^1.5.8",
     "moment": "^2.24.0",
     "prop-types": "^15.7.2",
-    "react": "16.8.3",
-    "react-native": "0.59.8",
-    "react-native-elements": "^1.1.0",
-    "react-native-gesture-handler": "^1.2.1",
-    "react-native-image-crop-picker": "^0.24.1",
+    "react": "16.9.0",
+    "react-native": "0.61.2",
+    "react-native-elements": "^1.2.6",
+    "react-native-gesture-handler": "^1.4.1",
+    "react-native-image-crop-picker": "^0.26.0",
+    "react-native-keep-awake": "^4.0.0",
     "react-native-keyboard-aware-scrollview": "^2.0.0",
-    "react-native-mail": "^3.0.7",
-    "react-native-modal": "^10.0.0",
-    "react-native-modal-datetime-picker": "^7.4.0",
-    "react-native-permissions": "^1.1.1",
-    "react-native-root-toast": "^3.0.2",
-    "react-native-status-bar-height": "^2.3.1",
-    "react-native-svg": "^9.4.0",
+    "react-native-mail": "^4.0.0",
+    "react-native-modal": "^11.4.0",
+    "react-native-modal-datetime-picker": "^7.6.0",
+    "react-native-permissions": "^2.0.2",
+    "react-native-root-toast": "^3.1.2",
+    "react-native-status-bar-height": "^2.4.0",
+    "react-native-svg": "^9.11.1",
     "react-native-timer": "^1.3.6",
-    "react-native-vector-icons": "^6.4.2",
-    "react-navigation": "^3.9.1",
-    "react-redux": "^7.0.3",
-    "reactotron-react-native": "^3.5.0",
-    "reactotron-redux": "^3.1.0",
+    "react-native-vector-icons": "^6.6.0",
+    "react-navigation": "^4.0.10",
+    "react-navigation-stack": "^1.9.4",
+    "react-redux": "^7.1.1",
+    "reactotron-react-native": "^4.0.2",
+    "reactotron-redux": "^3.1.2",
     "reactotron-redux-saga": "^4.2.2",
-    "realm": "^2.27.0-rc.3",
-    "redux": "^4.0.1",
-    "redux-saga": "^1.0.2",
+    "realm": "^3.2.0",
+    "redux": "^4.0.4",
+    "redux-saga": "^1.1.1",
     "reduxsauce": "^1.1.0",
-    "rn-fetch-blob": "^0.10.15",
+    "rn-fetch-blob": "^0.11.2",
     "seamless-immutable": "^7.1.4",
-    "styled-components": "^4.2.0",
-    "uuid": "^3.3.2",
+    "styled-components": "^4.4.0",
+    "uuid": "^3.3.3",
     "yup": "^0.27.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.4",
-    "@babel/runtime": "^7.4.4",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.8.0",
-    "babel-plugin-root-import": "^6.1.0",
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb": "^17.1.0",
+    "@babel/core": "^7.6.4",
+    "@babel/runtime": "^7.6.3",
+    "babel-eslint": "^10.0.3",
+    "babel-jest": "^24.9.0",
+    "babel-plugin-root-import": "^6.4.1",
+    "eslint": "^6.5.1",
     "eslint-import-resolver-babel-plugin-root-import": "^1.1.1",
-    "eslint-plugin-import": "^2.17.2",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-react": "^7.13.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-react": "^7.16.0",
     "eslint-plugin-react-native": "^3.7.0",
-    "jest": "^24.8.0",
-    "metro-react-native-babel-preset": "^0.54.0",
-    "react-test-renderer": "16.8.3"
+    "jest": "^24.9.0",
+    "metro-react-native-babel-preset": "^0.56.0",
+    "react-test-renderer": "16.9.0"
   },
   "jest": {
     "preset": "react-native"
