@@ -4,10 +4,10 @@
 
 import 'react-native';
 import React from 'react';
-import App from '../App';
+import renderer from 'react-test-renderer';
+import App from '../src';
 
 // Note: test renderer must be required after react-native.
-import renderer from 'react-test-renderer';
 
 it('renders correctly', () => {
   renderer.create(<App />);
