@@ -18,7 +18,7 @@ export const HeaderButton = styled.TouchableOpacity`
 
 export const HeaderIcon = styled(Icon).attrs({
   color: Colors.tintColor,
-  size: 24,
+  size: 30,
 })``;
 
 export const HeaderTitle = styled.Text`
