@@ -10,12 +10,20 @@ import Routes from '~/routes';
 
 const App = () => {
   // Dependencies Warnings
-  YellowBox.ignoreWarnings(['Warning: Async Storage has been extracted from react-native core']);
-  YellowBox.ignoreWarnings(['Require cycle: node_modules/rn-fetch-blob/index.js']);
+  YellowBox.ignoreWarnings([
+    'Warning: Async Storage has been extracted from react-native core',
+  ]);
+  YellowBox.ignoreWarnings([
+    'Require cycle: node_modules/rn-fetch-blob/index.js',
+  ]);
 
   return (
     <Provider store={store}>
-      <StatusBar backgroundColor="transparent" translucent barStyle="dark-content" />
+      <StatusBar
+        backgroundColor="transparent"
+        translucent
+        barStyle="dark-content"
+      />
       <Routes />
     </Provider>
   );
