@@ -1,20 +1,17 @@
 import styled from 'styled-components/native';
-import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scrollview';
 
 import Colors from '~/constants/Colors';
 
-export const Container = styled(KeyboardAwareScrollView).attrs({
-  enableOnAndroid: true,
-  contentContainerStyle: {
-    flex: 1,
-    width: '100%',
-    alignItems: 'center',
-  },
-})``;
+export const Container = styled.View`
+  flex: 1;
+  width: 100%;
+  align-items: center;
+`;
 
 export const AddButton = styled.TouchableOpacity`
   height: 44px;
-  background-color: ${props => (props.disabled ? Colors.lightGray : Colors.tintColor)};
+  background-color: ${props =>
+    props.disabled ? Colors.lightGray : Colors.tintColor};
   justify-content: center;
   align-items: center;
   padding: 0 20px;
