@@ -10,7 +10,8 @@ export const Container = styled.View`
 
 export const AddButton = styled.TouchableOpacity`
   height: 44px;
-  background-color: ${props => (props.disabled ? Colors.lightGray : Colors.tintColor)};
+  background-color: ${props =>
+    props.disabled ? Colors.lightGray : Colors.tintColor};
   justify-content: center;
   align-items: center;
   padding: 0 20px;
